Initial Commit
Basic implementation of a dictionary.  Cannot currently delete items,
but can insert and retrieve based on string key.  Able to either assign
a free_item function to the dictionary as a whole that is called on
every item in the dictionary when it is deleted, or assign a separate
deletion function to every item individually, in case the user wants to
have a dictionary with multiple types of values.  Currently, only able
to use linear probing for insertion.  Need to add quadratic probing, and
deletion.